requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6.3
click==8.1.7
Flask==2.2.5
Flask-Bcrypt==1.0.1
flask-cors==4.0.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
numpy==1.24.3
packaging==23.2
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==2.2.2
clarifai==9.8.1
langchain==0.0.295

